move unsigned values into each specific module

+++ core/src/zeroMain/kotlin/uk/tvidal/kpi/Gpio.kt
@@ -13,15 +13,15 @@ import uk.tvidal.kpi.gpio.GpioMode.OUTPUT
 import uk.tvidal.kpi.gpio.GpioValue.OFF
 import uk.tvidal.kpi.gpio.GpioValue.ON
 
-fun GpioPin.on() = bcm2835_gpio_set(pin)
-fun GpioPin.off() = bcm2835_gpio_clr(pin)
+fun GpioPin.on() = bcm2835_gpio_set(pin.toUByte())
+fun GpioPin.off() = bcm2835_gpio_clr(pin.toUByte())
 
-fun GpioPin.read() = when (bcm2835_gpio_lev(pin)) {
+fun GpioPin.read() = when (bcm2835_gpio_lev(pin.toUByte())) {
     LOW.toUByte() -> OFF
     else -> ON
 }
 
 fun GpioPin.setMode(mode: GpioMode) = when (mode) {
-    INPUT -> bcm2835_gpio_fsel(pin, BCM2835_GPIO_FSEL_INPT.toUByte())
-    OUTPUT -> bcm2835_gpio_fsel(pin, BCM2835_GPIO_FSEL_OUTP.toUByte())
+    INPUT -> bcm2835_gpio_fsel(pin.toUByte(), BCM2835_GPIO_FSEL_INPT.toUByte())
+    OUTPUT -> bcm2835_gpio_fsel(pin.toUByte(), BCM2835_GPIO_FSEL_OUTP.toUByte())
 }
